# docs/swagger.yaml
definitions:
...
  dto.VoluteerRequestCreatingDTO:
    properties:
      status:
        type: integer
      type:
        type: string
      user_id:
        type: integer
    required:
    - status
    - type
    - user_id
    type: object
info:
  contact: {}
  description: This is a volunteer service API
  title: Onboarding and Volunteer Service API
  version: "1.0"
paths:
...
  /api/v1/volunteer/request:
    post:
      consumes:
      - application/json
      description: Create a new volunteer request
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VoluteerRequestCreatingDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
      summary: Create a new volunteer request
      tags:
      - volunteer
securityDefinitions:
  bearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"